+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # LetsForm Changelog
 
+## Next
+- Added Mantine: tabs
+- Added RSuite5: color for checkboxes
+
 ## 0.7.6
 - Added JSON Schemas for all UI frameworks
 - Extracted react-hook-form from main bundle